Create django.yml
This commit introduces a new GitHub Actions workflow configuration specifically designed for continuous integration of a Django project. The workflow is triggered on any push or pull request to the 'development' branch. This ensures that all changes are automatically tested before they are merged, maintaining the integrity and reliability of the code base.